make serde_json optional
