+++ src/bson.rs
@@ -21,6 +21,9 @@
 
 //! BSON definition
 
+#[cfg(feature = "serde_json-1")]
+mod json;
+
 use std::{
     convert::TryFrom,
     fmt::{self, Debug, Display, Formatter},
@@ -28,10 +31,8 @@ use std::{
     ops::Index,
 };
 
-use serde_json::{json, Value};
-
 pub use crate::document::Document;
-use crate::{base64, oid, raw::CString, spec::ElementType, Binary, Decimal128};
+use crate::{oid, raw::CString, spec::ElementType, Binary, Decimal128};
 
 /// Possible BSON value types.
 #[derive(Clone, Default, PartialEq)]
@@ -450,143 +451,7 @@ where
     }
 }
 
-/// This will create the [relaxed Extended JSON v2](https://www.mongodb.com/docs/manual/reference/mongodb-extended-json/) representation of the provided [`Bson`](../enum.Bson.html).
-impl From<Bson> for Value {
-    fn from(bson: Bson) -> Self {
-        bson.into_relaxed_extjson()
-    }
-}
-
 impl Bson {
-    /// Converts the Bson value into its [relaxed extended JSON representation](https://www.mongodb.com/docs/manual/reference/mongodb-extended-json/).
-    pub fn into_relaxed_extjson(self) -> Value {
-        match self {
-            Bson::Double(v) if v.is_nan() => {
-                let s = if v.is_sign_negative() { "-NaN" } else { "NaN" };
-
-                json!({ "$numberDouble": s })
-            }
-            Bson::Double(v) if v.is_infinite() => {
-                let s = if v.is_sign_negative() {
-                    "-Infinity"
-                } else {
-                    "Infinity"
-                };
-
-                json!({ "$numberDouble": s })
-            }
-            Bson::Double(v) => json!(v),
-            Bson::String(v) => json!(v),
-            Bson::Array(v) => Value::Array(v.into_iter().map(Bson::into_relaxed_extjson).collect()),
-            Bson::Document(v) => Value::Object(
-                v.into_iter()
-                    .map(|(k, v)| (k, v.into_relaxed_extjson()))
-                    .collect(),
-            ),
-            Bson::Boolean(v) => json!(v),
-            Bson::Null => Value::Null,
-            Bson::RegularExpression(Regex { pattern, options }) => {
-                let mut chars: Vec<_> = options.as_str().chars().collect();
-                chars.sort_unstable();
-
-                let options: String = chars.into_iter().collect();
-
-                json!({
-                    "$regularExpression": {
-                        "pattern": pattern.into_string(),
-                        "options": options,
-                    }
-                })
-            }
-            Bson::JavaScriptCode(code) => json!({ "$code": code }),
-            Bson::JavaScriptCodeWithScope(JavaScriptCodeWithScope { code, scope }) => json!({
-                "$code": code,
-                "$scope": Bson::Document(scope).into_relaxed_extjson(),
-            }),
-            Bson::Int32(v) => v.into(),
-            Bson::Int64(v) => v.into(),
-            Bson::Timestamp(Timestamp { time, increment }) => json!({
-                "$timestamp": {
-                    "t": time,
-                    "i": increment,
-                }
-            }),
-            Bson::Binary(Binary { subtype, ref bytes }) => {
-                let tval: u8 = From::from(subtype);
-                json!({
-                    "$binary": {
-                        "base64": base64::encode(bytes),
-                        "subType": hex::encode([tval]),
-                    }
-                })
-            }
-            Bson::ObjectId(v) => json!({"$oid": v.to_hex()}),
-            Bson::DateTime(v) if v.timestamp_millis() >= 0 && v.to_time_0_3().year() <= 9999 => {
-                json!({
-                    // Unwrap safety: timestamps in the guarded range can always be formatted.
-                    "$date": v.try_to_rfc3339_string().unwrap(),
-                })
-            }
-            Bson::DateTime(v) => json!({
-                "$date": { "$numberLong": v.timestamp_millis().to_string() },
-            }),
-            Bson::Symbol(v) => json!({ "$symbol": v }),
-            Bson::Decimal128(v) => json!({ "$numberDecimal": v.to_string() }),
-            Bson::Undefined => json!({ "$undefined": true }),
-            Bson::MinKey => json!({ "$minKey": 1 }),
-            Bson::MaxKey => json!({ "$maxKey": 1 }),
-            Bson::DbPointer(DbPointer {
-                ref namespace,
-                ref id,
-            }) => json!({
-                "$dbPointer": {
-                    "$ref": namespace,
-                    "$id": {
-                        "$oid": id.to_hex()
-                    }
-                }
-            }),
-        }
-    }
-
-    /// Converts the Bson value into its [canonical extended JSON representation](https://www.mongodb.com/docs/manual/reference/mongodb-extended-json/).
-    pub fn into_canonical_extjson(self) -> Value {
-        match self {
-            Bson::Int32(i) => json!({ "$numberInt": i.to_string() }),
-            Bson::Int64(i) => json!({ "$numberLong": i.to_string() }),
-            Bson::Double(f) if f.is_normal() => {
-                let mut s = f.to_string();
-                if f.fract() == 0.0 {
-                    s.push_str(".0");
-                }
-
-                json!({ "$numberDouble": s })
-            }
-            Bson::Double(f) if f == 0.0 => {
-                let s = if f.is_sign_negative() { "-0.0" } else { "0.0" };
-
-                json!({ "$numberDouble": s })
-            }
-            Bson::DateTime(date) => {
-                json!({ "$date": { "$numberLong": date.timestamp_millis().to_string() } })
-            }
-            Bson::Array(arr) => {
-                Value::Array(arr.into_iter().map(Bson::into_canonical_extjson).collect())
-            }
-            Bson::Document(arr) => Value::Object(
-                arr.into_iter()
-                    .map(|(k, v)| (k, v.into_canonical_extjson()))
-                    .collect(),
-            ),
-            Bson::JavaScriptCodeWithScope(JavaScriptCodeWithScope { code, scope }) => json!({
-                "$code": code,
-                "$scope": Bson::Document(scope).into_canonical_extjson(),
-            }),
-
-            other => other.into_relaxed_extjson(),
-        }
-    }
-
     /// Get the [`ElementType`] of this value.
     pub fn element_type(&self) -> ElementType {
         match *self {
@@ -668,7 +533,7 @@ impl Bson {
                 } else {
                     doc! {
                         "$binary": {
-                            "base64": base64::encode(bytes),
+                            "base64": crate::base64::encode(bytes),
                             "subType": hex::encode([tval]),
                         }
                     }
