+++ src/extjson/models.rs
@@ -1,5 +1,7 @@
 //! A module defining serde models for the extended JSON representations of the various BSON types.
 
+#![cfg_attr(not(feature = "serde_json-1"), allow(unused))]
+
 use serde::{
     de::{Error as _, Unexpected},
     Deserialize,
@@ -196,6 +198,7 @@ impl Uuid {
     }
 }
 
+#[cfg(feature = "serde_json-1")]
 #[derive(Deserialize)]
 #[serde(deny_unknown_fields)]
 pub(crate) struct JavaScriptCodeWithScope {
